requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
azure-search-documents==11.4.0
openai==1.54.0
httpx==0.27.0
pytest==7.4.3
pytest-asyncio==0.21.1
pymongo==4.6.1
motor==3.3.2
